Fix harmless compiler warnings in FTS5.
FossilOrigin-Name: 29d9132a768e0460e22e41abb8b6188181ba0fc6c4665161d3d0af28d7583e19

+++ src/json.c
@@ -3453,7 +3453,7 @@ static char *jsonBadPathError(
   int rc                    /* Maybe JSON_LOOKUP_NOTARRAY */
 ){
   char *zMsg;
-  if( rc==JSON_LOOKUP_NOTARRAY ){
+  if( rc==(int)JSON_LOOKUP_NOTARRAY ){
     zMsg = sqlite3_mprintf("not an array element: %Q", zPath);
   }else{
     zMsg = sqlite3_mprintf("bad JSON path: %Q", zPath);
